Fix 16 bugs found in full codebase audit
Critical fixes:
- Fix affiliate commission using display fee (0 when absorbed) instead of actual platform fee
- Fix parseInt truncating decimal donations to parseFloat
- Fix PayPal refund sending NGN amount as USD without conversion
- Add idempotency check to Stripe webhook (skip already-completed orders)

High fixes:
- Fix Math.min() returning Infinity on empty ticket arrays
- Fix order.id.slice() crash when id is null
- Add processing->completed status flow to AdminProcessPayout with rollback
- Improve EventPayouts duplicate check with fallback query
- Fix NULL approval_status defaulting to 'approved' instead of 'pending'

Medium fixes:
- Fix date range filters using now instead of startOfDay for week/month
- Fix PlatformPnL silently dropping chargebacks in currencies with no orders
- Fix promoter payout float precision blocking exact balance payouts
- Fix fee rounding: sum rounded parts instead of re-rounding the sum
- Add email ownership check to ad checkout edge function
- Fix AdminFeeManagement preview NaN on empty inputs

Low fixes:
- Fix WebMyAds status logic for ads with no approval_status

+++ src/pages/WebFreeRSVP.jsx
@@ -287,7 +287,7 @@ export function WebFreeRSVP() {
   const acceptsDonations = event?.accepts_donations && event?.donation_amounts?.length > 0
   const donationOptions = event?.donation_amounts || []
   const allowCustomDonation = event?.allow_custom_donation !== false
-  const actualDonation = customDonation ? parseInt(customDonation) || 0 : donationAmount
+  const actualDonation = customDonation ? parseFloat(customDonation) || 0 : donationAmount
   
   // Donation fee handling - check if donor pays the fee or organizer absorbs it
   const donorPaysFee = event?.donation_fee_handling === 'pass_to_attendee'
